pyproject.toml
[project]
name = "casa-feita"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "JSkahl", email = "josekahl333@gmail.com" }]
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
    "Pillow>=10.4.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "[" }

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
[tool.pdm.dev-dependencies]
dev = ["isort>=5.13.2", "black>=24.8.0"]

[tool.black]
line-length = 120

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"